requirements.txt
openai>=2.26.0
python-dotenv>=1.0.0
requests>=2.31.0
python-frontmatter>=1.1.0
loguru>=0.7.0
flask>=3.0.0
